refactor: fix domain→ui layer violations surfaced by Konsist
DrawerUtils and OnboardingUtils imported from de.lemke.oneuisample.ui
while in the domain package. Move them to ui/. Extract Icon data class
from IconAdapter (ui.util) into domain/ so ObserveIconListUseCase can
reference it without a domain→ui import.
